requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
notion-client>=2.2.1
openai>=1.3.0
removebg>=0.4
PyPDF2>=3.0.0
Flask>=3.0.0
gunicorn>=21.2.0

